# api/openapi.yaml
openapi: 3.0.2
info:
  description: |
    PayGate is a RESTful API enabling first-party Automated Clearing House ([ACH](https://en.wikipedia.org/wiki/Automated_Clearing_House)) transfers to be created without a deep understanding of a full NACHA file specification. First-party transfers initiate at an Originating Depository Financial Institution (ODFI) and are sent off to other Financial Institutions.

    Tenants are the largest grouping in PayGate and are typically a vendor who is reselling ACH services or a company making ACH payments themselves. A legal entity is linked off a Tenant as the primary Customer used to KYC and in transfers with the Tenant itself.

    An Organization is a grouping within a Tenant which typically represents an entity making ACH transfers. These include clients of an ACH reseller or business accepting payments over ACH. A legal entity is linked off an Organization as the primary Customer used to KYC and in transfers with the Organization itself.

    ![](https://raw.githubusercontent.com/moov-io/paygate/master/docs/images/tenant-in-paygate.png)

    There are also [admin endpoints](https://moov-io.github.io/paygate/admin/) for back-office operations.
  version: v1
  title: Paygate API
  contact:
    url: https://github.com/moov-io/paygate
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: http://localhost:8082
    description: Local development
...
paths:
...
  /organizations:
    get:
      tags: [Organizations]
      summary: Get Organizations
      description: Retrieve all Organizations for the given tenantID
      operationId: getOrganizations
      parameters:
        - name: X-Request-ID
          in: header
          description: Optional requestID allows application developer to trace requests through the systems logs
          example: rs4f9915
          schema:
            type: string
        - name: X-Tenant
          in: header
          required: true
          description: Unique tenantID set by an auth proxy or client to identify and isolate objects.
          schema:
            type: string
      responses:
        '200':
          description: List of Organizations for the given tenantID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Organizations'
        '400':
          description: Problem getting Organizatios, see error
          content:
            application/json:
              schema:
                $ref: 'https://raw.githubusercontent.com/moov-io/api/master/openapi-common.yaml#/components/schemas/Error'
    post:
      tags: [Organizations]
      summary: Create Organization
      description: Create a new Organization under PayGate
      operationId: createOrganization
      parameters:
        - name: X-Request-ID
          in: header
          description: Optional requestID allows application developer to trace requests through the systems logs
          example: rs4f9915
          schema:
            type: string
        - name: X-Tenant
          in: header
          required: true
          description: Unique tenantID set by an auth proxy or client to identify and isolate objects.
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
              schema:
                $ref: '#/components/schemas/CreateOrganization'
      responses:
        '200':
          description: Created Organization successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Organization'
        '400':
          description: Problem creating Organization, see error
          content:
            application/json:
              schema:
                $ref: 'https://raw.githubusercontent.com/moov-io/api/master/openapi-common.yaml#/components/schemas/Error'
  /organizations/{organizationID}:
    put:
      tags: [Organizations]
      summary: Update Organization
      description: Update metadata for an Organization
      operationId: updateOrganization
      parameters:
        - name: organizationID
          in: path
          required: true
          description: organizationID for the Organization to update
          example: hs4f9470
          schema:
            type: string
        - name: X-Request-ID
          in: header
          description: Optional requestID allows application developer to trace requests through the systems logs
          example: rs4f9915
          schema:
            type: string
        - name: X-Tenant
          in: header
          required: true
          description: Unique tenantID set by an auth proxy or client to identify and isolate objects.
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
              schema:
                $ref: '#/components/schemas/CreateOrganization'
      responses:
        '200':
          description: Created Organization successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Organization'
        '400':
          description: Problem creating Organization, see error
          content:
            application/json:
              schema:
                $ref: 'https://raw.githubusercontent.com/moov-io/api/master/openapi-common.yaml#/components/schemas/Error'
  # Micro-Deposits
...
components:
  schemas:
...
    CreateOrganization:
      properties:
        name:
          type: string
          example: My Organization
          description: Legal name for this Organization
        tenantID:
          type: string
          example: 23cca67d
          description: tenantID to create this Organization under
        primaryCustomer:
          type: string
          example: 11ffa67d
          description: |
            A customerID from the Customers service to use in Transfers with this Organization. When transferring to or from the Organization this Customer and Account(s) are used.
            The Customer assigned here should represent the legal entity that manages the Organization.
      required:
        - name
        - tenantID
        - primaryCustomer
    Organization:
      properties:
        organizationID:
          type: string
          example: 45cdb67d
          description: Unique identifier for this Organization
        name:
          type: string
          example: My Organization
          description: Legal name for this Organization
        tenantID:
          type: string
          example: 23cca67d
          description: tenantID this Organization is created under
        primaryCustomer:
          type: string
          example: 11ffa67d
          description: |
            A customerID from the Customers service to use in Transfers with this Organization. When transferring to or from the Organization this Customer and Account(s) are used.
            The Customer assigned here should represent the legal entity that manages the Organization.
      required:
        - organizationID
        - name
        - tenantID
        - primaryCustomer
    Organizations:
      type: array
      items:
        $ref: '#/components/schemas/Organization'
...